requirements.txt
pillow >= 9.3.0
numpy >= 1.23.0
attrs >= 22.1.0
discord.py==1.7.3
python-dotenv
easygui